+++ codebase/libs/Mdv/src/NcfMdv/NcfMdvx.cc
@@ -168,7 +168,7 @@ int NcfMdvx::convertMdv2Ncf(const string &url)
     return -1;
   }
   stat_struct_t &fileStat = ncfFile.getStat();
-  off_t fileLen = fileStat.st_size;
+  size_t fileLen = fileStat.st_size;
   
   // read in buffer
